created auth providers folder and moved access-group-provider to the providers folder with updated import path
